requirements.txt
asgiref==3.2.7
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-rest-swagger==2.2.0
django-cors-headers==3.2.
drf-yasg==1.17.1
djoser==2.0.3
fcm-django==0.3.4
pyfcm==1.4.7
Markdown==3.2.1
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
beautifulsoup4==4.9.3
uWSGI==2.0.18